+++ packages/codemods/src/transforms/v2/web-react/__testfixtures__/tooltipmodern-component-name.input.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+// @ts-ignore: No declaration -- The library is not installed; we don't need to install it for fixtures.
+import { TooltipModern } from '@lmc-eu/spirit-web-react';
+
+export const MyComponent = () => (
+  <>
+    <TooltipModern>Tooltip</TooltipModern>
+  </>
+);
