+++ packages/codemods/src/transforms/v2/web-react/__tests__/tooltipmodern-component-name.test.ts
@@ -1,0 +1,3 @@
+import { testTransform } from '../../../../../tests/testUtils';
+
+testTransform(__dirname, 'tooltipmodern-component-name');
