+++ packages/web-react/src/components/Tooltip/Tooltip.tsx
@@ -1,36 +1,99 @@
-import classNames from 'classnames';
-import React, { LegacyRef, forwardRef, useMemo } from 'react';
+import React, { useRef } from 'react';
 import { useStyleProps } from '../../hooks';
 import { SpiritTooltipProps } from '../../types';
-import TooltipCloseButton from './TooltipCloseButton';
-import { useTooltipStyleProps } from './useTooltipStyleProps';
+import { TooltipProvider } from './TooltipContext';
+import { useFloating } from './useFloating';
 
-export const Tooltip = forwardRef((props: SpiritTooltipProps, ref) => {
-  const { children, isDismissible, closeLabel = 'Close', open, onClose, ...restProps } = props;
+const Tooltip = (props: SpiritTooltipProps) => {
+  const {
+    children,
+    enableFlipping: flipProp = true,
+    enableFlippingCrossAxis: flipCrossAxis = true,
+    enableShifting: shiftProp = true,
+    enableSizing: sizeProp = true,
+    flipFallbackAxisSideDirection = 'none',
+    flipFallbackPlacements = ['bottom', 'top'],
+    id,
+    isDismissible = false,
+    isFocusableOnHover = false,
+    isOpen = false,
+    onToggle,
+    placement: tooltipPlacement,
+    trigger = ['click', 'hover'],
+    ...rest
+  } = props;
 
-  const { classProps, props: modifiedProps } = useTooltipStyleProps({
-    isDismissible,
-    open,
-    ...restProps,
+  const { styleProps, props: otherProps } = useStyleProps({ ...rest });
+
+  // Refs for FloatingUI
+  const arrowRef = useRef<HTMLSpanElement>(null);
+  const tooltipRef = useRef<HTMLDivElement>(null);
+
+  // Get `--tooltip-max-width` and `--tooltip-offset` from CSS variables
+  let tooltipMaxWidth;
+  let tooltipOffset;
+  let tooltipCornerOffset;
+  let tooltipArrowWidth;
+  const tooltipElement = tooltipRef.current?.querySelector('[data-spirit-element="tooltip"]');
+  const tooltipArrowElement = tooltipElement?.querySelector('[data-spirit-element="tooltip-arrow"]');
+
+  if (tooltipElement) {
+    const tooltipComputedStyle = window.getComputedStyle(tooltipElement);
+    const tooltipArrowComputedStyle = tooltipArrowElement && window.getComputedStyle(tooltipArrowElement);
+    tooltipMaxWidth = parseInt(tooltipComputedStyle.getPropertyValue('--tooltip-max-width'), 10);
+    tooltipOffset = parseInt(tooltipComputedStyle.getPropertyValue('--tooltip-offset'), 10);
+    tooltipCornerOffset = tooltipArrowComputedStyle
+      ? parseInt(tooltipArrowComputedStyle.getPropertyValue('--tooltip-arrow-corner-offset'), 10)
+      : 0;
+    tooltipArrowWidth = tooltipArrowComputedStyle ? parseInt(tooltipArrowComputedStyle.width, 10) : 0;
+  }
+
+  // Get props for the FloatingUI hook
+  const { getFloatingProps, getReferenceProps, maxWidth, middlewareData, placement, refs, x, y } = useFloating({
+    arrowRef,
+    cornerOffset: tooltipCornerOffset,
+    flipCrossAxis,
+    flipFallbackAxisSideDirection,
+    flipFallbackPlacements,
+    flipProp,
+    isFocusableOnHover,
+    isOpen,
+    offset: tooltipOffset,
+    onToggle,
+    shiftProp,
+    sizeProp,
+    tooltipArrowWidth,
+    tooltipMaxWidth,
+    tooltipPlacement,
+    trigger,
   });
-  const { styleProps, props: transferProps } = useStyleProps({ ...modifiedProps });
-  const renderCloseButton = useMemo(
-    () => isDismissible && <TooltipCloseButton onClick={onClose} label={closeLabel} />,
-    [isDismissible, onClose, closeLabel],
-  );
 
   return (
-    <div
-      className={classNames(classProps.rootClassName, styleProps.className)}
-      style={styleProps.style}
-      ref={ref as LegacyRef<HTMLDivElement> | undefined}
-      {...transferProps}
+    <TooltipProvider
+      value={{
+        anchorRef: refs.setPositionReference,
+        arrowRef,
+        getFloatingProps,
+        getReferenceProps,
+        id,
+        isDismissible,
+        isOpen,
+        middlewareData,
+        onToggle,
+        placement,
+        sizeMaxWidth: maxWidth,
+        tooltipMaxWidth,
+        tooltipRef: refs.setFloating,
+        triggerRef: refs.setReference,
+        x,
+        y,
+      }}
     >
-      {children}
-      {renderCloseButton}
-      <span className={classProps.arrowClassName} />
-    </div>
+      <div ref={tooltipRef} {...styleProps} {...otherProps}>
+        {children}
+      </div>
+    </TooltipProvider>
   );
-});
+};
 
 export default Tooltip;
