+++ packages/web-react/src/components/Tooltip/demo/TooltipDismissible.tsx
@@ -1,21 +1,22 @@
 import React, { useState } from 'react';
 import { Button } from '../../Button';
-import TooltipWrapper from '../TooltipWrapper';
-import Tooltip from '../Tooltip';
+import { Tooltip, TooltipPopover, TooltipTrigger } from '..';
 
 const TooltipDismissible = () => {
   const [open, setOpen] = useState(true);
 
-  const openHandler = () => setOpen(true);
-  const closeHandler = () => setOpen(false);
-
   return (
-    <TooltipWrapper UNSAFE_className="d-inline-block">
-      <Button onClick={openHandler}>I have a tooltip 😎</Button>
-      <Tooltip isDismissible placement="right" open={open} onClose={closeHandler}>
-        Close me
-      </Tooltip>
-    </TooltipWrapper>
+    <Tooltip
+      id="TooltipDismissible"
+      isOpen={open}
+      onToggle={setOpen}
+      placement="right"
+      isDismissible
+      trigger={['click']}
+    >
+      <TooltipTrigger elementType={Button}>I have a tooltip 😎</TooltipTrigger>
+      <TooltipPopover>Close me</TooltipPopover>
+    </Tooltip>
   );
 };
