+++ packages/web-react/src/components/Tooltip/demo/TooltipDismissibleViaJS.tsx
@@ -1,14 +1,10 @@
 import React, { useEffect, useState } from 'react';
 import { Button } from '../../Button';
-import TooltipWrapper from '../TooltipWrapper';
-import Tooltip from '../Tooltip';
+import { Tooltip, TooltipPopover, TooltipTrigger } from '..';
 
 const TooltipDismissibleViaJS = () => {
   const localStorageValue = localStorage.getItem('my-tooltip-react') === 'true';
   const [open, setOpen] = useState(localStorageValue);
-
-  const openHandler = () => setOpen(true);
-  const closeHandler = () => setOpen(false);
 
   useEffect(() => {
     localStorage.setItem('my-tooltip-react', open.toString());
@@ -17,12 +13,17 @@ const TooltipDismissibleViaJS = () => {
   return (
     <>
       <p>Saves data to local storage.</p>
-      <TooltipWrapper UNSAFE_className="d-inline-block">
-        <Button onClick={openHandler}>I have a tooltip 😎</Button>
-        <Tooltip isDismissible placement="right" open={open} onClose={closeHandler}>
-          Close me
-        </Tooltip>
-      </TooltipWrapper>
+      <Tooltip
+        id="TooltipDismissibleViaJS"
+        isOpen={open}
+        onToggle={setOpen}
+        placement="right"
+        isDismissible
+        trigger={['click']}
+      >
+        <TooltipTrigger elementType={Button}>I have a tooltip 😎</TooltipTrigger>
+        <TooltipPopover>Close me</TooltipPopover>
+      </Tooltip>
     </>
   );
 };
