+++ packages/web-react/src/components/Tooltip/demo/TooltipHover.tsx
@@ -1,24 +1,28 @@
-import React from 'react';
-import { TooltipModern, TooltipPopover, TooltipTrigger } from '..';
+import React, { useState } from 'react';
 import { Button } from '../../Button';
+import { Tooltip, TooltipPopover, TooltipTrigger } from '..';
 
 const TooltipHover = () => {
-  const [open, setOpen] = React.useState(false);
+  const [open, setOpen] = useState(false);
 
   return (
-    <TooltipModern
+    <Tooltip
       id="tooltip-on-hover-floating-ui"
       isOpen={open}
       onToggle={setOpen}
       placement="right"
       flipFallbackPlacements={['left', 'top-start', 'bottom-start']}
+      trigger={['hover']}
+      isFocusableOnHover
     >
       <TooltipTrigger elementType={Button}>I have a tooltip 😎</TooltipTrigger>
       <TooltipPopover>
         Hello there! <br />
-        <a href="#">Click me</a>
+        <a href="#" className="link-inverted">
+          Click me
+        </a>
       </TooltipPopover>
-    </TooltipModern>
+    </Tooltip>
   );
 };
