+++ packages/web-react/src/components/Tooltip/demo/UncontrolledTooltip.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { Button } from '../../Button';
+import { UncontrolledTooltip, TooltipPopover, TooltipTrigger } from '..';
+
+const UncontrolledTooltipDemo = () => (
+  <UncontrolledTooltip id="UncontrolledTooltip" placement="left" isFocusableOnHover isDismissible>
+    <TooltipTrigger elementType={Button}>I have a tooltip 😎</TooltipTrigger>
+    <TooltipPopover>Close me</TooltipPopover>
+  </UncontrolledTooltip>
+);
+
+export default UncontrolledTooltipDemo;
