--- packages/web-react/src/components/TooltipModern/README.md
@@ -1,9 +1,0 @@
-# TooltipModern
-
-⚠️ `TooltipModern` component is [deprecated][deprecated] and will be renamed to `Tooltip` in the next major version.
-
-For more information and implementation examples, please visit the [`TooltipModern` section][tooltip-modern-section] section in the [readme][tooltip] for the `Tooltip` component.
-
-[deprecated]: https://github.com/lmc-eu/spirit-design-system/tree/main/packages/web-react/README.md#deprecations
-[tooltip-modern-section]: https://github.com/lmc-eu/spirit-design-system/blob/main/packages/web-react/src/components/Tooltip/README.md#tooltipmodern
-[tooltip]: https://github.com/lmc-eu/spirit-design-system/blob/main/packages/web-react/src/components/Tooltip/README.md
