--- packages/web-react/src/components/TooltipModern/demo/TooltipDismissible.tsx
@@ -1,25 +1,0 @@
-import React, { useState } from 'react';
-import { Button } from '../../Button';
-import { TooltipModern, TooltipPopover, TooltipTrigger } from '../../Tooltip';
-
-const TooltipDismissible = () => {
-  const [open, setOpen] = useState(true);
-
-  return (
-    <div className="docs-Stack docs-Stack-stretch">
-      <TooltipModern
-        id="TooltipDismissible"
-        isOpen={open}
-        onToggle={setOpen}
-        placement="right"
-        isDismissible
-        trigger={['click']}
-      >
-        <TooltipTrigger elementType={Button}>I have a tooltip 😎</TooltipTrigger>
-        <TooltipPopover>Close me</TooltipPopover>
-      </TooltipModern>
-    </div>
-  );
-};
-
-export default TooltipDismissible;
