--- packages/web-react/src/components/TooltipModern/index.html
@@ -1,1 +1,0 @@
-{{> web-react/demo}}
