--- packages/web-react/src/components/TooltipModern/index.ts
@@ -1,1 +1,0 @@
-export * from '../Tooltip';
